% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_CumCount.R
\name{AddGroupCol}
\alias{AddGroupCol}
\title{Handle all dfSubjects operations, only if GroupID is not yet in Numerator or Denominator frames.
Will also filter all events with no GroupID}
\usage{
AddGroupCol(df, dfSubjects, strSubjectCol, strGroupCol, strGroupLevel)
}
\description{
Handle all dfSubjects operations, only if GroupID is not yet in Numerator or Denominator frames.
Will also filter all events with no GroupID
}
\keyword{internal}
