% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtrapolateDenominator.R
\name{ExtrapolateDenominator}
\alias{ExtrapolateDenominator}
\title{Extrapolate Denominator Events}
\usage{
ExtrapolateDenominator(
  dfDenominator,
  dfNumerator,
  strSubjectCol,
  strDenominatorDateCol,
  strInstanceNameCol,
  vLikePatternInstanceName = c("\%unsch\%", "\%disc\%"),
  nMinSubjectRatioInstance = 0.7
)
}
\arguments{
\item{dfDenominator}{\code{data.frame} with a column for SubjectID and \code{strDenominatorDateCol}}

\item{dfNumerator}{\code{data.frame} with a column for SubjectID and \code{strNumeratorDateCol}}

\item{strSubjectCol}{\code{character} Column name in \code{dfSubjects} to use for subject ID.}

\item{strDenominatorDateCol}{\code{character} Column name in \code{dfDenominator} to use for denominator calculation.}

\item{strInstanceNameCol}{\code{character} When provided will extrapolate planned Numerator events e.g. visits. Use
for terminal binary events such as patient discontinuations to avoid survival bias. Default: NULL}

\item{vLikePatternInstanceName}{\verb{character vector} vector of sql like patterns to filter instance names for
extrapolation. Default: c("\%unsch\%", "\%disc\%")}

\item{nMinSubjectRatioInstance}{numeric, minimum subject ratio per instance name to consider for extrapolation.
Default: 0.7}
}
\description{
internal function called by \code{\link[=Input_CumCount]{Input_CumCount()}}
Filters all instances not matching vLikePatternInstanceName that occur
for a minimum ratio of subjects. Attempts to estimate the order of instances
from subject-level instance dates and to establish the median time difference
between instances. Missing instances for subjects with entries in numerator
data frame will be added using the median days between instances to extrapolate
dates. No instances will be added if extrapolated date exceeds the maximum instance
date for subjects without numerator entries.
}
