% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Simaerep.R
\name{Flag_Simaerep}
\alias{Flag_Simaerep}
\title{Flag Simaerep}
\usage{
Flag_Simaerep(dfAnalyzed, vThreshold)
}
\arguments{
\item{dfAnalyzed}{\code{data.frame} as returned by \code{\link[=Analyze_Simaerep]{Analyze_Simaerep()}}}

\item{vThreshold}{\code{numeric} vector of thresholds to flag groups. Must be between 0 and 1.}
}
\value{
\code{data.frame} with the following specification:\tabular{lll}{
   Column Name \tab Description \tab Type \cr
   GroupID \tab The group ID \tab Character \cr
   MetricExpected \tab Expected ratio from simulations \tab Numeric \cr
   MetricGroup \tab Ratio all subjects in GroupID \tab Numeric \cr
   OverReportingProbability \tab Probability over-reporting numerator events \tab Numeric \cr
   UnderReportingProbability \tab Probability under-reporting numerator events \tab Numeric \cr
   Flag \tab Flag for group \tab Numeric \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Flag over and under-reporting probabilities. The sum of under- and over reporting probability
per group is usually zero, however \code{\link[simaerep:simaerep]{simaerep::simaerep()}} adjusts edge cases in which group ratio
and expected ratio are all zero'
}
\examples{
# {clindata} Example for cumulative AE per Visit Count
dfInput <- Input_CumCount(
    dfSubjects = clindata::rawplus_dm,
    dfNumerator = clindata::rawplus_ae,
    dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
    strSubjectCol = "subjid",
    strGroupCol = "siteid",
    strGroupLevel = "Site",
    strNumeratorDateCol = "aest_dt",
    strDenominatorDateCol  = "visit_dt"
  )

dfAnalyzed <- Analyze_Simaerep(dfInput)
Flag_Simaerep(dfAnalyzed, vThreshold = c(0.95, 0.99))
}
\seealso{
\code{\link[=Analyze_Simaerep]{Analyze_Simaerep()}}
}
