% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.simaerep-package.R
\docType{package}
\name{gsm.simaerep-package}
\alias{gsm.simaerep}
\alias{gsm.simaerep-package}
\title{gsm.simaerep: Add simaerep KRI to gsm}
\description{
This is a simaerep extension for the gsm package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impala-consortium.github.io/gsm.simaerep/}
}

}
\author{
\strong{Maintainer}: Björn Koneswarakantha \email{bjoern.koneswarakantha@roche.com} (\href{https://orcid.org/0000-0003-4585-7799}{ORCID})

}
\keyword{internal}
