% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_BarChartSimaerep.R
\name{Widget_BarChartSimaerep}
\alias{Widget_BarChartSimaerep}
\title{Simawerep Bar Chart Widget}
\usage{
Widget_BarChartSimaerep(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "ExpectedNumerator",
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  strOutputLabel = paste0(fontawesome::fa("chart-simple", fill = "#337ab7"),
    "  Bar Chart"),
  vResultTooltipKeys = c("ExpectedNumerator", "Score", "Metric", "Numerator",
    "Denominator"),
  ...
)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{lMetric}{`list` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()] and
turing it into a list. Expected columns: `File`,`MetricID`, `Group`,
`Abbreviation`, `Metric`, `Numerator`, `Denominator`, `Model`, `Score`, and
`strThreshold`. For more details see the Data Model vignette:
`vignette("DataModel", package = "gsm.kri")`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{vThreshold}{\code{numeric} Threshold values.}

\item{strOutcome}{\code{character} Outcome variable. Default: 'ExpectedNumerator'.}

\item{bAddGroupSelect}{`logical` Add a dropdown to highlight sites? Default: `TRUE`.}

\item{strShinyGroupSelectID}{`character` Element ID of group select in Shiny context. Default: `'GroupID'`.}

\item{strOutputLabel}{`character` Label describing the output object, which is assigned to the
`output_label` attribute of the output object and appears in the report generated by
[gsm.kri::Report_KRI()]..}

\item{vResultTooltipKeys}{\code{character} Result tooltip keys.
Default: c("ExpectedNumerator", "Score", "Metric", "Numerator", "Denominator").}

\item{...}{`any` Additional chart configuration settings.}
}
\description{
modified \code{\link[gsm.kri:Widget_BarChart]{gsm.kri::Widget_BarChart()}} to include ExpectedNumerator as the default outcome.
}
\examples{
## Filter data to one metric and snapshot
reportingResults_filter <- gsm.core::reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

reportingMetrics_filter <- gsm.core::reportingMetrics \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001") \%>\%
  as.list()

## Make chart
Widget_BarChartSimaerep(
  dfResults = reportingResults_filter,
  dfGroups = gsm.core::reportingGroups,
  lMetric = reportingMetrics_filter,
  vThreshold = reportingMetrics_filter$Threshold
)

}
