% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Simaerep.R
\name{Analyze_Simaerep}
\alias{Analyze_Simaerep}
\title{Analyze Simaerep}
\usage{
Analyze_Simaerep(dfInput, r = 1000)
}
\arguments{
\item{dfInput}{\code{data.frame} as returned by \code{\link[=Input_CumCount]{Input_CumCount()}}}

\item{r}{Integer or tbl_object, number of repetitions for bootstrap
simulation. Pass a tbl object referring to a table with one column and as
many rows as desired repetitions. Default: 1000.}
}
\value{
\code{data.frame} with the following specification:\tabular{lll}{
   Column Name \tab Description \tab Type \cr
   GroupID \tab The group ID \tab Character \cr
   GroupLevel \tab The group level \tab Character \cr
   Numerator \tab Numerator events \tab Numeric \cr
   Denominator \tab Denominator events \tab Numeric \cr
   Metric \tab Ratio all subjects in GroupID \tab Numeric \cr
   OverReportingProbability \tab Probability over-reporting numerator events \tab Numeric \cr
   UnderReportingProbability \tab Probability under-reporting numerator events \tab Numeric \cr
   Score \tab Combined Score between \tab Numeric \cr
   ExpectedNumerator \tab Count of Expected Numerator Events \tab Numeric \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate over and under-reporting probabilities for numerator denominator ratio using \code{\link[simaerep:simaerep]{simaerep::simaerep()}}.

Uses inframe \code{simaerep} method and does not apply multiplicity correction as per latest recommendations.

The sum of under- and over reporting probability  per group is usually zero, however \code{\link[simaerep:simaerep]{simaerep::simaerep()}}
adjusts edge cases in which group ratio and expected ratio are all zero. We combine both metrics into a score
between 0 and 1, where 0 is under-reporting and 1 is over-reporting.
}
\examples{
# {clindata} Example for cumulative AE per Visit Count
dfInput <- Input_CumCount(
  dfSubjects = clindata::rawplus_dm,
  dfNumerator = clindata::rawplus_ae,
  dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
  strSubjectCol = "subjid",
  strGroupCol = "siteid",
  strGroupLevel = "Site",
  strNumeratorDateCol = "aest_dt",
  strDenominatorDateCol = "visit_dt"
)

Analyze_Simaerep(dfInput)

}
\seealso{
\code{\link[simaerep:simaerep]{simaerep::simaerep()}}, \code{\link[=Input_CumCount]{Input_CumCount()}}
}
