% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Simaerep.R
\name{Flag_Simaerep}
\alias{Flag_Simaerep}
\title{Flag_Simaerep}
\usage{
Flag_Simaerep(...)
}
\arguments{
\item{...}{arguments passed to \code{gsm.core::Flag()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Alias for \code{gsm.core::Flag()}
}
\seealso{
\code{\link[gsm.core:Flag]{gsm.core::Flag()}}
}
