% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Simaerep.R
\name{Flag_Simaerep}
\alias{Flag_Simaerep}
\title{Flag_Simaerep}
\usage{
Flag_Simaerep(dfAnalyzed, ...)
}
\arguments{
\item{dfAnalyzed}{\code{data.frame} where flags should be added.}

\item{...}{parameters passed to \code{\link[gsm:Flag]{gsm::Flag()}}}
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Wrapper for \code{gsm::Flag()}
}
\seealso{
\code{\link[gsm:Flag]{gsm::Flag()}}
}
