% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_CumCount.R
\name{Input_CumCount}
\alias{Input_CumCount}
\title{Cumulative Numerator Event Count per Denominator Event Count}
\usage{
Input_CumCount(
  dfSubjects,
  dfNumerator,
  dfDenominator,
  strGroupCol,
  strGroupLevel = NULL,
  strSubjectCol,
  strNumeratorCol = NULL,
  strDenominatorCol = NULL,
  strNumeratorDateCol,
  strDenominatorDateCol,
  strOrphanedMethod = c("filter", "assign")
)
}
\arguments{
\item{dfSubjects}{\code{data.frame} with columns for SubjectID and any other relevant subject information}

\item{dfNumerator}{\code{data.frame} with a column for SubjectID and \code{strNumeratorDateCol}}

\item{dfDenominator}{\code{data.frame} with a column for SubjectID and \code{strDenominatorDateCol}}

\item{strGroupCol}{\code{character} Column name in \code{dfSubjects} to use for grouping.}

\item{strGroupLevel}{\code{character} value for the group level. Default: NULL defaults to \code{strGroupCol}}

\item{strSubjectCol}{\code{character} Column name in \code{dfSubjects} to use for subject ID.}

\item{strNumeratorCol}{\code{character} Column name in \code{dfNumerator} to use for numerator ID. Default: NULL}

\item{strDenominatorCol}{\code{character} Column name in \code{dfDenominator} to use for denominator ID. Default: NULL}

\item{strNumeratorDateCol}{\code{character} Column name in \code{dfNumerator} to use for numerator calculation.}

\item{strDenominatorDateCol}{\code{character} Column name in \code{dfDenominator} to use for denominator calculation.}

\item{strOrphanedMethod}{\code{character} one_of("filter", "assign") filter orphaned numerator events or assign
to random patient enrolled at site the time of event. Default: "filter"}
}
\value{
\code{data.frame} with the following specification:\tabular{lll}{
   Column Name \tab Description \tab Type \cr
   SubjectID \tab The subject ID \tab Character \cr
   GroupID \tab The group ID \tab Character \cr
   GroupLevel \tab The group type \tab Character \cr
   Numerator \tab Cumulative Count Numerator \tab Numeric \cr
   Denominator \tab Cumulative Count Denominator \tab Numeric \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate a subject level cumulative numerator event count per denominator event count.

This function takes in a list of data frames including dfSUBJ, dfNumerator, and dfDenominator,
and calculates a subject level cumulative count.

The data requirements for the function are as follows:
\itemize{
\item dfSubjects: A data frame with columns for SubjectID and any other relevant subject information
\item dfNumerator: A data frame with a column for SubjectID and \code{strNumeratorCol} if \code{strNumeratorMethod} is "Sum"
\item dfDenominator: A data frame with a column for SubjectID and \code{strDenominatorCol} if \code{strDenominatorMethod} is "Sum"
}

All other columns are dropped from the output data frame. Note that if no values for a subject are
found in dfNumerator/dfDenominator numerator and denominator values are filled with 0 in the output data frame.
}
\examples{
# Run for AE KRI
Input_CumCount(
    dfSubjects = clindata::rawplus_dm,
    dfNumerator = clindata::rawplus_ae,
    dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
    strSubjectCol = "subjid",
    strGroupCol = "siteid",
    strGroupLevel = "Site",
    strNumeratorDateCol = "aest_dt",
    strDenominatorDateCol  = "visit_dt"
  )

}
\keyword{internal}
