% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric_Simaerep.R
\name{Visualize_Metric_Simaerep}
\alias{Visualize_Metric_Simaerep}
\title{Visualize_Metric Function}
\usage{
Visualize_Metric_Simaerep(
  dfResults,
  dfInput,
  dfMetrics = NULL,
  dfGroups = NULL,
  strMetricID = NULL,
  strSnapshotDate = NULL,
  bDebug = FALSE,
  vColors = c(`0` = "#9ED782", `1` = "#FEAA01", `2` = "#FF5858", `-1` = "#FEAA01", `-2` =
    "#FF5858", `NA` = "#a9a9a9"),
  ...
)
}
\arguments{
\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfInput}{data.frame created by \code{\link[=Input_CumCount]{Input_CumCount()}}}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm.core")`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{strMetricID}{`character` MetricID to subset the data.}

\item{strSnapshotDate}{`character` Snapshot date to subset the data.}

\item{bDebug}{`logical` Display console in html viewer for debugging. Default is `FALSE`.}

\item{vColors}{vector, named hex values for every Flag value in dfFlagged$Flag, Default NULL}

\item{...}{Additional chart configuration settings.}
}
\value{
A list containing the following charts:
\itemize{
\item simaerep: A simaerep plot using JavaScript.
\item scatterPlot: A scatter plot using JavaScript.
\item barChart: A bar chart using JavaScript with metric on the y-axis.
\item timeSeries: A time series chart using JavaScript with score on the y-axis.
\item metricTable: A table containing all
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function creates all available charts for a metric using the data provided
}
\examples{

 dfInput <- Input_CumCount(
   dfSubjects = clindata::rawplus_dm,
   dfNumerator = clindata::rawplus_ae,
   dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
   strSubjectCol = "subjid",
   strGroupCol = "invid",
   strGroupLevel = "Site",
   strNumeratorDateCol = "aest_dt",
   strDenominatorDateCol = "visit_dt"
 )

 dfAnalyzed <- Analyze_Simaerep(dfInput)
 dfFlagged <- Flag_Simaerep(dfAnalyzed, vThreshold = c(-0.99, -0.95, 0.95, 0.99))

Visualize_Metric_Simaerep(
  dfResults = dfFlagged,
  dfInput = dfInput
)

}
