% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Simaerep.R
\name{Visualize_Simaerep}
\alias{Visualize_Simaerep}
\title{Visualize Simaerep}
\usage{
Visualize_Simaerep(
  dfInput,
  dfFlagged,
  strStudyId = "StudyID",
  strScoreCol = "Score",
  nSiteMax = 16,
  vColors = NULL,
  strDenominator = "Denominator",
  strNumerator = "Numerator"
)
}
\arguments{
\item{dfInput}{data.frame created by \code{\link[=Input_CumCount]{Input_CumCount()}}}

\item{dfFlagged}{data.frame created by \code{\link[=Flag_Simaerep]{Flag_Simaerep()}}}

\item{strStudyId}{character, study label, Default: "StudyID"}

\item{strScoreCol}{character, name of score column in dfFlagged, Default: "Score"}

\item{nSiteMax}{integer, maximum of flagged sites to plot, Default: 16}

\item{vColors}{vector, named hex values for every Flag value in dfFlagged$Flag, Default NULL}

\item{strDenominator}{vector, label for Denominator x-column, Default: "Denominator"}

\item{strNumerator}{vector, label for Numerator y-column, Default: "Numerator"}
}
\description{
create a ggplot2 visualisation for a simaerep KRI
}
\details{
A widget that creates a simaerep visualisation of group-level metric results.
It plots the mean cumulative numerator count per denominator in the left panel and
highlights groups based on the over and under-reporting probability calculated
by the simaerep bootstrap algorithm. Flagged groups are shown in the right panel
including the total numerator counts per single patient.
}
\examples{
dfInput <- Input_CumCount(
  dfSubjects = clindata::rawplus_dm,
  dfNumerator = clindata::rawplus_ae,
  dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
  strSubjectCol = "subjid",
  strGroupCol = "invid",
  strGroupLevel = "Site",
  strNumeratorDateCol = "aest_dt",
  strDenominatorDateCol = "visit_dt"
)

dfAnalyzed <- Analyze_Simaerep(dfInput)
dfFlagged <- Flag_Simaerep(dfAnalyzed, vThreshold = c(-0.99, -0.95, 0.95, 0.99))

Visualize_Simaerep(dfInput, dfFlagged)
}
