% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_Simaerep.R
\name{Widget_Simaerep}
\alias{Widget_Simaerep}
\title{Simaerep Widget}
\usage{
Widget_Simaerep(
  dfInput,
  dfFlagged,
  dfGroups = NULL,
  lMetric = NULL,
  strStudyId = "StudyID",
  strScoreCol = "Score",
  vColors = NULL,
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  strOutputLabel = paste0(fontawesome::fa("chart-line", fill = "#337ab7"), "  Simaerep"),
  ...
)
}
\arguments{
\item{dfInput}{data.frame created by \code{\link[=Input_CumCount]{Input_CumCount()}}}

\item{dfFlagged}{data.frame created by \code{\link[=Flag_Simaerep]{Flag_Simaerep()}}}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{lMetric}{`list` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()] and
turing it into a list. Expected columns: `File`,`MetricID`, `Group`,
`Abbreviation`, `Metric`, `Numerator`, `Denominator`, `Model`, `Score`, and
`strThreshold`. For more details see the Data Model vignette:
`vignette("DataModel", package = "gsm.kri")`.}

\item{strStudyId}{character, study label, Default: "StudyID"}

\item{strScoreCol}{character, name of score column in dfFlagged, Default: "Score"}

\item{vColors}{vector, named hex values for every Flag value in dfFlagged$Flag, Default NULL}

\item{bAddGroupSelect}{`logical` Add a dropdown to highlight sites? Default: `TRUE`.}

\item{strShinyGroupSelectID}{`character` Element ID of group select in Shiny context. Default: `'GroupID'`.}

\item{strOutputLabel}{`character` Label describing the output object, which is assigned to the
`output_label` attribute of the output object and appears in the report generated by
[gsm.kri::Report_KRI()]..}

\item{...}{`any` Additional chart configuration settings.}
}
\value{
An htmlwidget object that can be rendered in R Markdown reports or Shiny apps.
}
\description{
A widget that creates a simaerep visualisation of group-level metric results.
It plots the mean cumulative numerator count per denominator in the left panel and
highlights groups based on the over and under-reporting probability calculated
by the simaerep bootstrap algorithm. Flagged groups are shown in the right panel
including the total numerator counts per single patient.
}
\examples{

 dfInput <- Input_CumCount(
   dfSubjects = clindata::rawplus_dm,
   dfNumerator = clindata::rawplus_ae,
   dfDenominator = clindata::rawplus_visdt \%>\% dplyr::mutate(visit_dt = lubridate::ymd(visit_dt)),
   strSubjectCol = "subjid",
   strGroupCol = "invid",
   strGroupLevel = "Site",
   strNumeratorDateCol = "aest_dt",
   strDenominatorDateCol = "visit_dt"
 )

 dfAnalyzed <- Analyze_Simaerep(dfInput)
 dfFlagged <- Flag_Simaerep(dfAnalyzed, vThreshold = c(-0.99, -0.95, 0.95, 0.99))

 Widget_Simaerep(dfInput, dfFlagged)

}
\seealso{
\code{\link{Widget_SimaerepOutput}} for use in Shiny apps
\code{\link{renderWidget_Simaerep}} for use in Shiny apps
}
